+++ src/Exceptions/InvalidApplyWithCallbackException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Khalyomede\ReorderBeforeAfter\Exceptions;
+
+use Exception;
+
+final class InvalidApplyWithCallbackException extends Exception
+{
+}