+++ tests/Misc/Product.php
@@ -4,12 +4,13 @@ declare(strict_types=1);
 
 namespace Tests\Misc;
 
-final readonly class Product
+final class Product
 {
     public function __construct(
-        public string $name,
-        public int $quantity,
-        public float $unitPrice,
+        public readonly string $name,
+        public readonly int $quantity,
+        public readonly float $unitPrice,
+        public int $order,
     ) {
     }
 }