+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Ability to create the listing from an array ([#3](https://github.com/khalyomede/reorder-before-after/issues/3)).
 - Ability to get all items from a listing ([#2](https://github.com/khalyomede/reorder-before-after/issues/2)).
+- Ability to set a callback to apply the order on your value ([#1](https://github.com/khalyomede/reorder-before-after/issues/10)).
 
 ## [0.1.0] - 2023-01-11